Add gun raycast shooting

+++ FPS/Assets/Scripts/Interfaces/IDamageable.cs
@@ -2,6 +2,6 @@ namespace Assets.Scripts.Interfaces
 {
     public interface IDamageable
     {
-        void Damage(int amount);
+        void Damage(float amount);
     }
 }
